create new integration/source folder and factor common code into helper.go
