+++ pkg/reconciler/integration/source/resources/containersource_test.go
@@ -101,56 +101,3 @@ func TestNewContainerSource(t *testing.T) {
 		t.Errorf("NewContainerSource() mismatch (-want +got):\n%s", diff)
 	}
 }
-
-func TestGenerateEnvVarsFromStruct(t *testing.T) {
-	type TestStruct struct {
-		Field1 int    `json:"field1"`
-		Field2 bool   `json:"field2"`
-		Field3 string `json:"field3"`
-	}
-
-	prefix := "TEST_PREFIX"
-	input := &TestStruct{
-		Field1: 123,
-		Field2: true,
-		Field3: "hello",
-	}
-
-	// Expected environment variables including SSL settings
-	want := []corev1.EnvVar{
-		{Name: "TEST_PREFIX_FIELD1", Value: "123"},
-		{Name: "TEST_PREFIX_FIELD2", Value: "true"},
-		{Name: "TEST_PREFIX_FIELD3", Value: "hello"},
-	}
-
-	got := generateEnvVarsFromStruct(prefix, input)
-
-	if diff := cmp.Diff(want, got); diff != "" {
-		t.Errorf("generateEnvVarsFromStruct() mismatch (-want +got):\n%s", diff)
-	}
-}
-
-func TestGenerateEnvVarsFromStruct_S3WithCamelTag(t *testing.T) {
-	type AWSS3 struct {
-		Arn    string `json:"arn,omitempty" camel:"CAMEL_KAMELET_AWS_S3_SOURCE_BUCKETNAMEORARN"`
-		Region string `json:"region,omitempty"`
-	}
-
-	prefix := "CAMEL_KAMELET_AWS_S3_SOURCE"
-	input := AWSS3{
-		Arn:    "arn:aws:s3:::example-bucket",
-		Region: "us-west-2",
-	}
-
-	// Expected environment variables including SSL settings and camel tag for Arn
-	want := []corev1.EnvVar{
-		{Name: "CAMEL_KAMELET_AWS_S3_SOURCE_BUCKETNAMEORARN", Value: "arn:aws:s3:::example-bucket"},
-		{Name: "CAMEL_KAMELET_AWS_S3_SOURCE_REGION", Value: "us-west-2"},
-	}
-
-	got := generateEnvVarsFromStruct(prefix, input)
-
-	if diff := cmp.Diff(want, got); diff != "" {
-		t.Errorf("generateEnvVarsFromStruct_S3WithCamelTag() mismatch (-want +got):\n%s", diff)
-	}
-}
